# showtime/showtime-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Showtime API
  version: 1.0.0
  description: Basketball club management backend API.
servers:
  - url: http://localhost:8080
    description: Local
security:
  - sessionAuth: []
...
paths:
...
  /api/v1/teams/generate:
    post:
      tags: [Team]
      operationId: generateTeams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamGenerateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamGenerateResponse'
  /api/v1/teams/generate-from-attendance:
    post:
      tags: [Team]
      operationId: generateTeamsFromAttendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamFromAttendanceRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamGenerateResponse'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
...
  schemas:
...
    SkillLevel:
      type: string
      enum: [LOW, LOWER_MID, MID, UPPER_MID, HIGH]
    Position:
      type: string
      enum: [GUARD, FORWARD, CENTER]
    PlayStyle:
      type: string
      enum: [SHOOTER, SLASHER, PLAYMAKER, DEFENDER, REBOUNDER, BALANCED]
...
    MemberProfile:
      type: object
      required: [id, name, heightCm, skillLevel, position, style]
      properties:
        id:
          type: string
        name:
          type: string
        heightCm:
          type: integer
          minimum: 130
          maximum: 240
        skillLevel:
          $ref: '#/components/schemas/SkillLevel'
        position:
          $ref: '#/components/schemas/Position'
        style:
          $ref: '#/components/schemas/PlayStyle'
    TeamGenerateRequest:
      type: object
      required: [teamNames, attendees]
      properties:
        teamNames:
          type: array
          minItems: 1
          items:
            type: string
        attendees:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/MemberProfile'
        previousTeamByMemberId:
          type: object
          additionalProperties:
            type: string
    TeamFromAttendanceRequest:
      type: object
      required: [attendanceDate, teamNames]
      properties:
        attendanceDate:
          type: string
          format: date
        teamNames:
          type: array
          minItems: 1
          items:
            type: string
        previousTeamByMemberId:
          type: object
          additionalProperties:
            type: string
    TeamResult:
      type: object
      required: [name, members, totalSkillScore, averageHeightCm, positionCounts, repeatedFromLastWeekCount]
      properties:
        name:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/MemberProfile'
        totalSkillScore:
          type: integer
        averageHeightCm:
          type: number
          format: double
        positionCounts:
          type: object
          additionalProperties:
            type: integer
        repeatedFromLastWeekCount:
          type: integer
    BalanceSummary:
      type: object
      required: [teamSkillGap, teamSizeGap, repeatedAssignments]
      properties:
        teamSkillGap:
          type: integer
        teamSizeGap:
          type: integer
        repeatedAssignments:
          type: integer
    TeamGenerateResponse:
      type: object
      required: [teams, summary]
      properties:
        teams:
          type: array
          items:
            $ref: '#/components/schemas/TeamResult'
        summary:
          $ref: '#/components/schemas/BalanceSummary'
...